Font width is CGFloat
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3945 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/AddressView/ADPersonPropertyView+Private.m
@@ -75,7 +75,7 @@
 {
   NSString *str;
   NSFont *font;
-  int w;
+  CGFloat w;
 
   str = ADLocalizedPropertyOrLabel(label);
   font = [self boldFont];
@@ -114,7 +114,8 @@
 				      details: (id) details
 {
   id cell;
-  NSSize cellSize; NSPoint cellOrigin;
+  NSSize cellSize;
+  NSPoint cellOrigin;
 
   cell = [[[ADPersonActionCell alloc] init] autorelease];
   switch(t)
@@ -693,7 +694,7 @@
 
 - (void) layout
 {
-  float heightBefore, heightAfter;
+  CGFloat heightBefore, heightAfter;
 
   _neededLabelWidth = 0;
